refactor: auth guard as separate file
