+++ lib/router/browser-router.js
@@ -1,35 +1,13 @@
+import authGuard from "./auth-guard";
+
 import { NotFound } from "../../src/pages/not-found";
-import AuthService from "../services/auth";
-import { findRouteByPath } from "../utils";
+
 export function createBrowserRouter(routes) {
   if (!Array.isArray(routes)) throw new Error("Provide routes as array");
 
   let currentCleanup = null;
 
-  const authGuard = (renderCallback) => {
-    return () => {
-      const currentPath = window.location.pathname;
-      let redirectTo = findRouteByPath(currentPath, routes);
-
-      const isLoggedIn = AuthService.isAuthenticated;
-
-      if (redirectTo) {
-        if (currentPath === "/login" && isLoggedIn) {
-          window.history.pushState({}, "", "/");
-          redirectTo = findRouteByPath("/", routes);
-        }
-
-        if (currentPath !== "/" && currentPath !== "/login" && !isLoggedIn) {
-          window.history.pushState({}, "", "/login");
-          redirectTo = findRouteByPath("/login", routes);
-        }
-      }
-
-      renderCallback({ redirectTo });
-    };
-  };
-
-  const render = authGuard(({ redirectTo }) => {
+  const _render = authGuard(routes, ({ redirectTo }) => {
     if (typeof currentCleanup === "function") {
       currentCleanup();
       currentCleanup = null;
@@ -47,12 +25,12 @@ export function createBrowserRouter(routes) {
 
   const navigate = (newPath) => {
     window.history.pushState({}, "", newPath);
-    render();
+    _render();
   };
 
   const run = () => {
-    window.addEventListener("popstate", render);
-    render();
+    window.addEventListener("popstate", _render);
+    _render();
   };
 
   return {
